+++ examples/dts-generation/src/pages/Home/models/counter.ts
@@ -1,0 +1,5 @@
+export default {
+  state: {
+    count: 1
+  }
+};