+++ packages/plugin-store/src/index.ts
@@ -79,9 +79,12 @@ export default async (api: any) => {
     routesPaths = [routes.routesPath];
   }
 
-  // add vite plugin for redirect page component
   if (vite) {
-    modifyUserConfig('vite.plugins', [vitePluginPageRedirect(rootDir, routesPaths)], { deepmerge: true });
+    modifyUserConfig(
+      'vite.plugins', 
+      [vitePluginPageRedirect(rootDir, routesPaths)],
+      { deepmerge: true }
+    );
   }
 
   if (swc) {
@@ -90,7 +93,7 @@ export default async (api: any) => {
         .rule('replace-router-path')
         // ensure that replace-router-path-loader is before babel-loader
         // @loadable/babel-plugin will transform the router paths which replace-router-path-loader couldn't transform
-        .before('babel-loader')
+        .after('tsx')
         .test((filePath: string) => routesPaths.includes(filePath))
         .use('replace-router-path-loader')
         .loader(require.resolve(path.join(__dirname, 'replacePathLoader')))