+++ packages/plugin-mpa/CHANGELOG.md
@@ -1,5 +1,13 @@
 # changelog
 
+## 2.0.2
+
+- [fix] compatible with the situation when config both `disableRuntime` and `mpa`
+
+## 2.0.1
+
+- [feat] support disable rewrite rules for FaaS render
+
 ## 2.0.0
 
 - [feat] upgrade build-scripts